[이가은] vo.*Img 필드(imgSize) 타입 변경

+++ src/main/java/goodee/gdj58/shop_c/vo/CustomerImg.java
@@ -8,5 +8,5 @@ public class CustomerImg { // 고객 사진
 	private String customerImgSaveName; // 이미지 파일 이름
 	private String customerImgOriginName; // 이미지 원본 파일 이름
 	private String customerImgKind; // 파일 타입
-	private int customerImgSize; // 파일 용량
+	private Long customerImgSize; // 파일 용량
 }
